VirtuaDoc
VirtuaDoc is a medical web-based application that helps users connect to physicians in real-time, schedule and set medical appointments, and lookup complex prescription information. To view VirtuaDoc you'll need to download the application Figma which is a free web-based application, once downloaded you can inspect the code and view it from there. This code is protected by the GNU General Public License. Be kind to each other and treat others as you would want to be treated, that has nothing to do with this application it's just good life advice :).